refactor(#3464): extract voice-announcement scope guard to satisfy frozen ratchet
The voice-announcement scoping fix (drop unauthorized announcements so non-owning
agents don't reply → multi-agent reply storm) added ~96 lines to intake_turn.rs,
a LoC-frozen giant (baseline 3771), breaking the giant_file_ratchet (3807>3771).

Extract the decision (`should_drop_unauthorized_voice_announcement`) + a
`drop_unauthorized_voice_announcement` wrapper (decision + the one-shot warn) +
the 4 unit tests into a new `voice_announcement_scope.rs`, leaving intake_turn.rs
with a single `if drop_unauthorized_voice_announcement(...) { return Ok(()); }`
call site. Behaviour is byte-identical to the PR (same drop decision, same two
warns, same return-Ok short-circuit). intake_turn.rs drops to 3766 (< 3771);
change-surfaces.md synced.
