+++ docs/agent-maintenance/change-surfaces.md
@@ -725,9 +725,11 @@
     force-clean watcher-respawn follow-through + always-run cross-tick
     retry/dead-man (P1-a: no early return on zero candidates), delegating the
     new behaviour to `health/watcher_respawn.rs`).
-  - `src/services/discord/router/message_handler/intake_turn.rs` (3769 lines;
+  - `src/services/discord/router/message_handler/intake_turn.rs` (3766 lines;
     Discord message intake turn orchestration split from the router message
-    handler; bugfix only outside a further extraction plan; +9 from #3082
+    handler; bugfix only outside a further extraction plan; #3464 extracted the
+    unauthorized-voice-announcement scope decision to `voice_announcement_scope.rs`;
+    +9 from #3082
     queued-only answer-flush gate (`is_queued_notice` on the two
     `send_intake_placeholder` call sites: `true` for the race-lost queued card,
     `false` for the active-turn placeholder); +57 from #3182 normal-dequeue
